chore(index.html): update social media icons with correct image sources
The social media icons in the index.html file have been updated with the correct image sources. The LinkedIn icon now points to "./images/linkedin-ico.png", the GitHub icon now points to "./images/github-ico.png", and the Instagram icon now points to "./images/insta-ico.png". This change ensures that the correct icons are displayed on the website.