feat(index.html): add new section for displaying projects with links to code and live versions
The changes in this commit add a new section to the index.html file. This section is called "Projects" and it includes a heading with the text "Projects" and a subheading with a description. The subheading explains that all the projects listed in this section will have links to the code and live versions. This change was made to enhance the user experience and provide easy access to the projects showcased on the website.