feat(index.html): add image and update project details for Snake Game
The Snake Game project has been added to the projects section of the website. An image of the game has been included to provide a visual representation. The project details have been updated to provide information about the project, including the fact that it was built for the "Founders and Coders" software development apprenticeship. The project was built using only HTML, CSS, and JavaScript. A link to a blog post about the project has been added for further information.